Add boilerplate codde for mysql HA support

+++ dbaas/workflow/steps/util/restore_snapshot/remove_data.py
@@ -31,19 +31,20 @@ class RemoveData(BaseStep):
             if return_code != 0:
                 raise Exception(str(output))
 
-            for host in workflow_dict['not_primary_hosts']:
-                cs_host_attr = CsHostAttr.objects.get(host=host)
-                command = 'rm -rf /data/data/*'
+            if workflow_dict['not_primary_hosts'] > 1:
+                for host in workflow_dict['not_primary_hosts']:
+                    cs_host_attr = CsHostAttr.objects.get(host=host)
+                    command = 'rm -rf /data/data/*'
 
-                output = {}
-                return_code = exec_remote_command(server=host.address,
-                                                  username=cs_host_attr.vm_user,
-                                                  password=cs_host_attr.vm_password,
-                                                  command=command,
-                                                  output=output)
+                    output = {}
+                    return_code = exec_remote_command(server=host.address,
+                                                      username=cs_host_attr.vm_user,
+                                                      password=cs_host_attr.vm_password,
+                                                      command=command,
+                                                      output=output)
 
-                if return_code != 0:
-                    raise Exception(str(output))
+                    if return_code != 0:
+                        raise Exception(str(output))
 
             return True
         except Exception:
